requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.1
kiwisolver==1.4.4
matplotlib==3.7.2
networkx==3.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pip==23.2.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scipy==1.11.1
setuptools==65.5.0
six==1.16.0
tzdata==2023.3